requirements.txt
python-fasthtml
jieba>=0.42.1 